📝 Add MrklSystemPlanner and remove IterativePlanner
This commit adds a new example for using the MrklSystemPlanner, which is
a planning extension that uses the MRKL system to generate plans for
natural language goals. The example shows how to use the planner with
various skills, such as web search, language calculator, and time.

This commit also removes the IterativePlanner example, which is no longer
needed, as the MrklSystemPlanner is more robust and flexible. The
IterativePlanner extension is also removed from the project references.

Additionally, this commit adds a reference to the NCalcSkills project,
which provides a skill for evaluating mathematical expressions. This
skill is used in the MrklSystemPlanner example to compute the age of
Leo DiCaprio's girlfriend raised to the 0.43 power.
